Update Main component
imported the comments, partners, and promotions modules:

updated the constructor in Main component to add the imported arrays into the state:

used the filter array method to filter for objects where the featured property evaluates as true. 

Then, because filter returns an array and you want to pass an object, you will use [0] to access the first (and only) object in the array.